+++ handshake-nodejs-reporters/packages/handshake-dashboard/backup/core/TestEntity/constants.ts
@@ -1,4 +1,0 @@
-export const childBadge = 'This suite has a parent suite';
-export const retriedBadge =
-    'Does not contribute to the aggregated values as we have same suite which was retried recently.';
-export const parentBadge = "This suite has no parent, it's a top-level suite.";