+++ handshake-nodejs-reporters/packages/handshake-dashboard/src/pages/api/Import/Runs/[...id].ts
@@ -1,7 +1,6 @@
 import { readFileSync } from 'node:fs';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { join } from 'node:path';
-import { setTimeout as sleep } from 'node:timers/promises';
 
 type ResponseData = {
     message: string;
@@ -13,8 +12,6 @@ export default async function handler(
 ) {
     const { id } = request.query;
 
-    await sleep(3000); // 😴 [3 seconds]
-
     const filePath = join(
         process.env.TEST_RESULTS ?? '',
         'Import',
@@ -22,6 +19,5 @@ export default async function handler(
     );
 
     const value = JSON.parse(readFileSync(filePath, { encoding: 'utf8' }));
-    await sleep;
     response.status(200).json(value);
 }