+++ handshake-nodejs-reporters/packages/common-handshakes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # common-handshakes
 
+## 0.8.0
+
+### Minor Changes
+
+- ebdc7ae: packages are available under "@handshake/" and we are now using pnpm instead of npm
+
 ## 0.7.0
 
 ### Minor Changes